More Delicious Documentation
Wrote some more long comments to describe what's going on. Hopefully
this stuff is easily understandable. I guess I could test it on some
programming friends to see how good of a job I'm doing. I'm actually
quite proud of this work, I don't usually document stuff this well.

I also added some checks for null pointers and a check in the primary
query function to exit if a value is out of the range of the table.