Exhaustive Nullifiability Test
Since we can't say the equivalent sets method will find everything, we
need to comb over anything it doesn't find with just an exhaustive test
for nullifiability. That's precisely what this is. It's just a recursive
function that will check for nullifiability before doing every single
possible arithmetic operation and then checking that as a new set.