Set Tree Query Outputs to Function
Instead of printing to a file stream, performing a query on a set tree
requires passing in a function pointer to have the sets passed into when
found. This way, more manipulation can be done with the actual set, as
opposed to the query being the final step before printing. This is so
that we can have that final step of testing remaining sets after 1-M
equivalent pairs does its work.