This Program Should Work
Yep! It should just work now! I've put in the exhaustive test, and I've
tested it on a couple known ranges, nothing special as I want to wait to
make this discovery with my friend tomorrow morning. Gosh, I can't wait!
But yeah, the only other things I need to do for this version 1 is
polish up the documentation, and maybe tweak the way the program outputs
stuff...