Node Flagging Returns Current Status
I want to optimize the Innullifiability program, and I think the biggest
thing is making sure that the program doesn't go through the same thing
multiple times with regards to equivalent sets. There are many different
ways to get to a set, and so it would be great if the program could
notice it's already marked a set and not bother generating all the
equivalent for it as well and re-marking them as well yadda yadda yadda.

So here I've changed the Node Flagging function to return 1 in the case
that all the nodes have already been flagged on the tree, meaning the
set has already been marked. I'll use this to stop the Equivalent Sets
program from continuing to generate on a set that's it's already done.