Patched Node Flagging, Updated Test Program
I apparently just forgot that recursive call of `nodeFlag` even existed.
I wasn't checking the return code and I didn't notice it when trying to
add this new functionality with checking if nodes are already flagged.

Also, I just forgot I had a test program it seems. All the enum values
were wrong. But now I'm using it to test whether this new feature works
and it seems to be fine.