HOLY CRAP IT'S SO MUCH FASTER
WWWWWWOOOOOOWWWWWW!!!!!! I kinda expected it to be a lot faster, but
maybe not this fast... it takes 1,073 ms on my system as opposed to like
35,000 before this optimization. It's so simple too. It's iNCrEdIBlE!

If it wasn't clear already, yes, I've implemented this new feature in
the main program. Basically I just changed the void function parameter
in the Equivalent Sets program to a boolean function, which can tell it
when to stop generating more sets.