Re-Organized Tree Allocation Functions
I figure that I ought to optimize memory usage for the program by
allowing the tree to be dynamically allocated and deallocated. So I'm
re-labelling the existing recursive node (de)allocation functions as
operating on 'descendant nodes,' and I'm going to create a new function
to allocate 'child nodes,' as in direct children. This will allow for
the tree to only allocate new nodes when it needs.

So this will be implemented just in the Node Flagging function. Whenever
it needs to go to a lower-level node, it will just create the node's
children right there. Their descendants don't need to be recursively
generated as the function will recurse and allocate them if need be. In
addition, if a node is flagged, all its descendants can be deallocated
immediately. Not just the children, all of them. So yeah, should be
fairly simple to implement after I write the child allocation function.