Descendant Allocation Functions
The allocation functions that claim to operate on 'descendants' now
always take in a node and operate solely on its descendants. The
allocate function now takes in a node and just recursively allocates the
descendants, and doesn't allocate a new node to return. The free
function now recursively allocates the descendants but not the given
node itself. The root node needs to be manually allocated and freed now.

Also, I think this whole time I was never freeing the memory allocated
to hold arrays of pointers to child nodes. I was just freeing the node
structure without freeing the dynamically allocated memory for the
array. But now it should be better. I really should run some analysis
tools on this program at some point, there certainly are some more bugs.