Tree Query Function Independence
If the tree is to be dynamically allocated, the query function needs to
be able to handle itself when nodes aren't allocated. When it's looking
for marked sets, it switches into 'everything' mode when it finds a
flagged node, so I used that fact to have it completely ignore whatever
node was passed to it in the case where it's in that mode, as it's just
enumerating all the sets algorithmically anyways. This should mean it's
safe to use when working with a tree where flagged nodes have their
descendants deallocated.

I've tested it on the same normal statically allocated trees and it
seems to be fine, I see no reason why it wouldn't work on a dynamically
allocated tree.