Tree Dynamic Allocation Option
Implemented an option for a tree to be dynamically allocated. The
allocation mode will be passed in to construct the tree and will be used
by the Node Flagging function for it to know when to (de)allocate nodes.
For now, the only real difference is that the tree wouldn't be allocated
upon construction. There is no deallocation upon flagging yet.

Dynamically allocated trees don't completely work yet, the test is
failing. Patch is coming right up.