Dynamic Tree Allocation Works
The Query function wasn't accounting for the possibility that a node may
not have allocated children and yet not be flagged. This could happen if
some number simply is never marked. Also, I changed the free function so
that a null pointer as the child node array indicates that child nodes
aren't allocated.

Anyways, at this point I'm pretty sure dynamic tree allocation is
working properly. Next I'll implement freeing up memory instead of just
allocating it when needed.