Polished Tree Library & Main Program
Changed some of the type names in the Tree Library to make it clear that
they're for the tree. I'm going to be phasing out the tree and instead
using a different data structure. Also finished up a bit of
documentation.

In the main program, I implemented a little `-s` option for statically
allocating the tree, as it might be important for the program to execute
quickly. Yeah, I'm getting rid of the tree, but I still think it'd be
nice to have. I also changed some of the strings and implemented a
couple extra input checks.

This version of the program is looking pretty good, and I think I'll
make a little tag on it to save it before I go ahead and change a bunch
of stuff and possibly break it for a while.