Basic Files for New Set Record Library
The new library I'm making is going to be called 'Set Record'. It's just
an array with elements corresponding to each set. The sets will be
ordered lexicographically, and so the array can be indexed using that
'Combinatorial Number System' you learned about on English Wikipedia.

I'm switching to this just because the trees get expensive as they get
larger. There isn't really anything I can do to optimize the memory much
more. I mean, to be fair, I can't really get an exact idea of how much
memory it's using since `free()` won't actually return the memory to the
system and the system monitor might just be showing a bunch of stuff
that the program owns but isn't using. Dunno.

But I want to optimize the program more by having it enumerate sets
non-recursively. See, the equivalent sets program finds a length-3
nullifiable set, and then it'll immediately work on expanding it as many
ways as it can, all the way down to N. It would be much better if I
could eliminate everything as fast as possible by having it find all the
length-3 sets first, and only once that's finished move on to expanding
those. I just don't see an easy way to do that with the tree without
implementing some whole new data structure or just using a whole bunch
of them.

I'm not sure how much I expect this to enable me to optimize the
program's runtime. But I do know that simply implementing this will make
the program significantly lighter on memory usage. Like, A LOT. An
N = 6, M = 40 tree costs 186 M, while this should only cost 3.84 M for
the same sets. Even if there aren't many additional time savings I can
implement, this will enable me to do that N = 7, M = 64 computation on
my computer.