Documentation for Mark Function
Added one of those long comments. It's actually a very nice function,
but I'm afraid of forgetting how it works or it being too unclear for
others to read. I do still need to write documentation for the library
as a whole, just to give a bit of an overview of the combinatorics
involved in indexing an array like this.