Changed Set Tree Mark Function Return Values
I'm going back to the tree library because I feel like it still might be
quite useful. The optimization I'm going to make with doing the smaller
sets first would make the tree consume a whole lot less memory, and at
that point it's possible for it to be a viable option for really large
M-value calculations where a statically allocated array might be too
large. Maybe.

What I did here was swap the return values of the mark function to make
it a bit more consistent with the way the record library's mark function
is working. The record library returns the number of newly marked sets,
so this function now returns 1 if there was at least one newly marked
set, and 0 if not. Getting an exact number is too much work and
unnecessary for this purpose.