Optimized Set Tree Node Children
The set tree nodes used a pointer to an array of pointers to child
nodes. I simplified this to a pointer to an array of child nodes. All
nodes are the same, so this works just as well. Those pointers are quite
expensive, making this a far better design decision.

As a result, the tree uses less than half the memory it used to. It also
makes the main program execute way faster, probably on account of cache
misses.

I think I theoretically might even be able to run the N = 7, M = 64 case
on my machine with the program in this state, but I want to do that
nullifiable set generation optimization first.