Integrating Set Record Library into Main Program
Rewrote the main program a bit so that the main data structure for
marking sets could be either a tree or an array record. I had to change
some stuff like the enums to let both libraries be included at once.

Also, it turns out the M choose N function was just implemented
incorrectly the entire time, and yet it worked on a small scale. But
that's fixed now, I'm pretty sure. I went through and tested it with set
values.

It works fine, but something strange is happening when using the array
record: sometimes some extra sets get through the equivalent series
program unmarked that don't get through using the tree. I undid the
previous optimization where if no sets were newly marked, it wouldn't
continue to expand the nullifiable set, and that seemed to prevent the
extra sets from getting through, so I suspect there might be an issue
with the mark function and its ability to keep track of newly marked
sets.