Memory Leaks, I figured out what's going on
Here I patched a couple of memory leaks I just never noticed, and made
the code a bit more specific in a couple places.

So, I discovered earlier that when using the array record for the
primary data structure for sets, some extra sets would slip through the
equivalent sets phase. I thought this was something to do with the array
record being implemented incorrectly, but there's nothing wrong with it
that I can see. What's going on here is a particular effect I'd
theorized about but never really encountered...

The idea behind the optimization I did previously was this: the
equivalent sets program might generate the same set multiple times
through different ways, and a set shouldn't have to be recursively
expanded multiple times. How does the program know if it's already done
a set? The way I implemented it was that the main program would use the
tree library to check if the set's already been marked everywhere. Now,
what if the set has been marked everywhere, but not by virtue of itself?
In other words, what if all nodes corresponding to a particular set just
happen to be marked? In that case, maybe it's a mistake to say that we
shouldn't look at expanding this set further; maybe there's still
another set to be eliminated.

But when I did this optimization, nothing really happened. The number of
sets that made it past the equivalent sets phase stayed the same. So I
figured that this effect might just be impossible or something. But now
it's clear that this effect is real, though the consequences of it might
not be that bad.

The reason it wasn't appearing with the tree is because of one simple
fact that I ignored: the tree still has nodes for subsets, and flagging
one of those is still counted as a mark. The tree, in addition to
storing the state of each set, also keeps track of (most of) the subsets
that are generated as well, and won't (usually) behave as though it's
seen a particular subset when it hasn't and all its supersets just
happen to be marked anyways. It manages to (mostly) get around this
effect. (The exceptions are subsets which have values too high for their
position, like how you can't have the maximum value at any point except
the last position in full sets.)

So this effect appears with the array records, at least in the program's
current state. It's not as if it makes a huge difference right now, it's
still only a couple sets extra that need to be exhaustively checked at
the end. Again, the next optimization I'm going to be making is having
the program generate equivalent sets of one size before going to larger
sizes. That'll be done by making an array record for each new set size.
I figure I could have it so that supersets of already-generated
nullifiable sets are ignored when making a new generation.

It's late and I'm not fully thinking through everything, but what I
thought of was having the array record hold multiple addressable bits so
that one could be used for marking sets as supersets of smaller
nullifiable sets, and one could be used simply for marking a set as
being a new nullifiable set to expand. I feel like that'd avoid the
effect while still being super fast. Just an idea.