Set Array Records now store Addressable Bits
I want to be able to hold more information for each set, in particular I
want to be able to mark sets as supersets of a nullifiable set, without
marking them as new nullifiable sets themselves to be expanded later. So
the set array record now provides the ability to mark a set by setting
or de-setting individual bits in a char, and query by particular bit
settings.

Next I'm going to update the test program and the main program so they
may compile.