Set Array Record Querying OR
There wasn't a way to query sets that have at least one bit set, only if
some particular bits were all set in a particular way. So, I've
repurposed the all-zeros bitmask for some special cases. It treats the
settings parameter as a new bitmask, and matches all sets that have at
least one of the bits specified there. But, if that's all zeros too, it
treats it as a wildcard and matches every set, since there isn't much
use in querying for nothing.