Incremental Set Length Expansion Optimization
I finally did it. After two garbage attempts I've finally found a way
that works well. So I'll go over what this optimization does again:
previously, equivalent set generation was recursive, and whenever a set
was generated, it would immediately recurse down and start expanding off
of that set. This does cover everything, but it tends to be really
inefficient and spends a lot of time at the bottom with the
largest-length sets. I've taken out that recursive element, and made it
so that set expansion only expands by one element at a time, and so once
ALL the length-3 nullifiable sets are generated, those all get expanded
into level-4 sets, and so on.

The great thing about this is that it eliminates a lot of sets early on,
as smaller sets have many more supersets than larger ones. And
importantly, we only really need to expand nullifiable sets which ARE
NOT supersets of smaller nullifiable sets, as we know that smaller set
has already been expanded, and we can't get anything new. So we can skip
all those and save a bunch of time.

I've implemented this using the bit system for the Set Array Records. I
have one bit that gets set when it gets generated as a nullifiable set,
and another that is set when it's a superset of a nullifiable set.
Wherever that superset bit is set, we know that the set doesn't need to
be expanded further.

So going forward, I have a couple things I need to work on. First, I
should rework the documentation for the main program, add a bit to the
Equivalent Sets program, and just create it for the Set Record library.
In addition to that, I should implement a system which keeps count of
the number of sets remaining, as I want to be able to see how it behaves
and if it might be worth it to just interrupt the program at some point
to skip to the manual checking phase, possibly through a signal. Next, I
should look into optimizing or memoizing the manual check, maybe
implementing it as a separate program entirely, or maybe even writing it
in Haskell. Finally, I want to multithread the program. I figure it can
be done with some reworking, and once I have that, I could run it on
larger computers.