Optimized Out some Redundant Marking
The program shouldn't have to deal with marking supersets of a set if it
knows it's already been over a set previously. I already knew this, and
I made the program skip marking supersets if a set already was marked as
nullifiable. But I wasn't setting that flag when marking supersets
themselves, so there was some redundant marking going on when the
program reached a set it'd only marked as a superset. This optimizes
that away by letting all supersets have the nullifiable flag as well.