Beginning Iterative Query Function
So, my idea for multithreading was to have a query routine that could be
run in parallel and guarantee complete coverage. Basically, I wanted
something that I could call N times for N threads, and it would check
the status of every Nth set, so I could call it with starting offsets
from 0 to N - 1.

In order to implement that though, I need to redo the query function
entirely, as the recursive implementation doesn't really work like that.
I'm going to be switching it over to iteration. I've created a function
that can take a set and replace it with the Nth set after it
lexicographically. I've tested the function pretty well and I think it
works. So I just need to do the actual iterative query, and write a new
wrapper function.