Completed Iterative Query Function
I finished the function for a complete query done iteratively. All I
needed to do was do a while-loop with the same check as before and two
lines to go to the next set. Now the wrapper function uses the iterative
query too, and it runs just about as fast as the recursive query. Now I
should be able to implement multithreaded query.