Set Up Thread Safety for Set Records
Now each instance of a Set Record has a Mutex Lock attached to it, used
to prevent simultaneous marking, which could cause race conditions. The
mark function will grab the lock when it wants to actually mark sets, so
as to prevent another possibly parallel call from overwriting stuff and
getting bits mixed up, which could be problematic. There's only one lock
for each record instance. This should make the Set Record library
Thread-Safe as it is.