Basic Multithreading in Main Program, Not Perfect
I've implemented multithreading in the main program! It's not fully
working right now, it only seems to work with two threads before it
starts to segfault, but it seems to be doing fine. It makes it run
considerably faster on two threads.

What I've done is create a new command-line parameter for the number of
threads to create. I've made a function for automatically creating the
threads for set querying and generation, as well as the function for the
entry point of the threads.

What I want to do is figure out why it isn't working for more than two
threads, and then after I've dealt with that, make sort-of the same
thing for set checking at the end. Actually, it'd be better if I made
the new function more general, more of a "Multithreaded Query" function
like I was thinking for the library, so that I could just use the same
function for any queries/traversals.