Refactored as General Threaded Query Function
I don't know why I didn't do this in the first place. So that function
for making a bunch of threads for set generation is now just a general
function for querying a set record with multiple threads. I had to add a
few parameters, but yeah. It seems to be working just fine, of course
unless you want more than two threads. I could now just go ahead and
make the verification phase use this too.