Multithreaded Query for Checking Phase
Now the exhaustive test phase is also multithreaded! Stuff is very quick
now on two cores; it's not completely halved, and indeed there is some
extra user time, which I think is just because of the mutex trying to
prevent simultaneous marking (even though it's highly unlikely something
will be trying to mark at the same exact spot--we can optimize that
later). One thing I kinda forgot about was the return value for query,
which isn't being dealt with at all in the multithreaded implementation.
I should take care of that.