Progress Indication
Added print calls for printing the number of sets remaining after each
generation. It's really cool to see the number of sets tick down. It
might be interesting also to examine possibly printing the number of
sets to be expanded for each generation, showing the amount of work that
needs to be done.

I could possibly make the program handle a signal during the generation
phase to go ahead and move on to the next phase, which might be helpful
if there are just so few sets remaining. Or maybe the program could
automatically detect whether it would be more work to go through with
another generation than to just move to the check phase.