Replaced Mutex Locks with C11 Atomic Operations
I wanted to try and speed up the program by making a whole array of
mutex locks where each one represented a portion of the record, but I
tried removing the lock entirely and realized that there were basically
no issues without it. Though it's possible, two threads writing to the
same exact spot seems very rare. But I re-discovered the addition of
atomic operations in C11, and decided to incorporate those just for
safety.

So now the elements in the record array itself are all atomic chars. I
read from and write to them exclusively using the atomic operations.
This should work pretty well. The only issue I think is that there could
be two parallel mark calls which mark some of the same elements with
different settings, but that scenario is pretty messed up anyways.

So I think I'll change it up a bit and maybe even remove the ability to
unset a bit, maybe just have a bitmask and have it OR the bits every
time. That's really all I need. That could make the program run way
faster, as it'll remove a lot of stuff from the main marking loop. I'll
do that, and I also still need to remove the mutex stuff that remains.

This sped up the program by a lot. I ran N = 7, M = 85 with 6 threads
like yesterday and it took over an hour, but it finished in under half
an hour with this change.