Mark Function only ORs Bits
I never really needed to unset bits, so I figured I might as well remove
the ability to as it sorta lets it possibly mess itself up. Now, the
mark function will only set bits using a single atomic OR operation.
Just provide a mask, and it'll OR it to the bits. No more second 'bits'
parameter for marking. The counts will always be accurate; there is no
possibility for race conditions. Also it's marginally faster, maybe.