Simplified Main Program
The main program now only deals with one generation at a time. It only
has two Set Records allocated at once, for the source and destination.
Instead of supersets being marked immediately on all further levels,
they are now marked by a first pass of the source record.

This simpler program flow should make it easier to implement exporting
to files as well as buffered records. The runtime is slightly longer,
but by only like 30% or something. It's not that bad, and it still
scales well with multithreading.

The progress checking needs to be reworked. It doesn't make sense as it
is; I'll have it print out stuff like number of new nullifiable sets, or
maybe I'll have a separate thread to update the totals every so often.