Set Record Import/Export Functions
I've implemented the file operations for set records, in the form of an
'import' function and an 'export' function. The files use a plaintext
header for conveying the N and M values, and then contain the raw binary
data from the array starting 4K into the file. I decided to have the
functions deal only in writing to the files, so they take in a pointer
to a file handler as opposed to a filename to open. Let the main program
deal with opening files.

I think I previously mentioned that "performance slowed by about 30%"
due to the simplification of the main program. That was completely wrong
and only based on the N = 6, M = 32 case. For larger cases the
difference is more marginal. So it's fine.

So, next up will be 'buffered records', or perhaps instead just
'segmentation'. I think I might make it as basic as possible: it'll just
be that a record is instantiated with a max block size, which is a
number of 4K pages it's allowed to take up, as well as an offset for
determining how many blocks are 'behind' it, so it knows effectively the
range of sets covered. I could simply run the program multiple times to
generate multiple output files, each one containing a segment. I could
also run multiple instances of the program, perhaps, which might be
ideal for a dual-processor machine to reduce the need for sharing
memory. All in all, that might enable me to find the first N = 8 set.

I've decided after all this time I spent freaking out about the build
system that I'd rather just write everything in one source file for each
program/library. Maybe I might end up '#include'ing individual functions
from separate files, but I don't want to deal with individual
compilation and linking. Seems weird. Don't want to expose hidden stuff.