Exporting Results from Main Program
Implemented exporting results from the main program through a filename
parameter. Thread count is now required. Also, I simplified the code for
generation by making a separate function for outputting the trivial
size-2 nullifiable sets and integrating the size-2 expansion into the
same for-loop as the rest. Still needs to be multithreaded, though.