Importing Records to Main Program
Implemented the ability to base generation(s) off of an input record,
instead of re-generating already-computed nullifiable sets. The input
record needs to have the same M-value as the program is working with.
The N-value needs to be manually passed in to the program. Importing and
exporting records enables the program to be used to compute nullifiable
sets of all sizes for a given M-value by iteratively running it with the
previous result as input.

Here's a basic bullet list of the remaining stuff I can think of that I
want to implement (I might be forgetting something):

- Program Invocation Options
    - Skip Exhaustive Test (for research into ones that can't be caught,
      or just to save execution time)
    - Don't Output Sets at the End (maybe have just a separate program
      to print out sets from a record)

- Set Record Segmentation (having only a particular segment of the full
  record loaded, just a range of sets; we don't have infinite memory)

- Set Record M-value Contiguity (simply: order sets by highest value
  instead of lowest, so an M-value's sets are all contiguous; could save
  memory if you know a minimum M-value by just only loading a range of
  M-values)

- Progress Checking
    - I think just have the Query function take in a pointer to a number
      that it updates with however many sets it's done
    - This could be monitored by a separate thread that updates progress
      bars or just numbers

- Multithread the Base Generation

- Try Optimizing the Exhaustive Test (if we're running it every
  generation, the very first computation must go above M)