Set Address Calculator
In preparation for changing the order system used by the Set Record, and
just to get rid of the ridiculously verbose recursive implementation of
the mark function, I'm re-implementing it iteratively using a function
for addressing a set individually. This iterative implementation would
loop across all values and insert each one into the set of constraints,
giving all the supersets. I guess it'd recurse too if it needs to expand
more than once. I have a feeling though that this implementation could
be much faster than the recursive one, as that one produces so many
stackframes.