Iterative Mark Implementation
Re-implemented the Mark function iteratively. It just takes an array of
the constraints and then inserts all the values it can. Yes, it does use
recursion, but not nearly as much. It only recurses to mark or further
expand a superset. I've tested it and it seems to work, but it's unclear
as to whether it's actually faster or slower. I'm gonna try on some more
hardware as I'm doing this from my laptop.