Switched to Highest-Value Combinadics Ordering
I've swapped in the iterative Mark function, and swapped in the new set
increment and index/set conversion functions where used. Set Records now
use this new order, whereby sets are laid out in ascending order by
their highest values (see 'Combinatorial Number System' on Wikipedia).
This could enable for better segmentation, where only certain M-values
are allocated to a record.

With this change, it seems that expansion has been substantially sped
up. Consistently over various cases, it seems to be running nearly 3x
faster. Woweee!!!

So next I'm going to probably just remove all the now-unused functions
and rename the ones I want to use. I might play around with the idea of
removing the superset logic from the library and just require every mark
call provide a full-length set, and just put the superset logic
somewhere else. However, it might be a good idea to keep it in and make
it smarter for segmentation. If a segment is restricted to certain
M-values, the superset calculator could just make sure one of them is
always included, and then it won't have a bunch of wasted calls for sets
not in the segment. I'm sure there are other things possible. But, of
course, if you're doing that, you might as well do something similar in
Equivalent Sets, where an expansion must always provide a certain
M-value. So yeah, maybe all that stuff could be extracted elsewhere and
just have extra bits for optimizing for segmentation.