Barebones Supersets Program
This will be the program that holds the logic to be extracted from the
Mark function in Set Records. It'll be very similar to the Equivalent
Sets program, in that it'll be expanding sets for marking, and it'll
have only two user-mode functions: initialization, and set expansion.
Right now they're just barebones functions in the actual source.