Main Program Superset Expansion
Switched the Main program to using the Supersets program, instead of
relying on the Set Records library for doing supersets. Now the whole
logic in Set Records for marking supersets is unnecessary, and it can be
removed to simplify the library.

Now the Main program refers to two modes of 'expansion': Supersets and
Mutation. Whereas before 'expansion' referred to calling on Equivalent
Sets, there is now a second program that expands sets (makes a
compatible set with one added element), so it makes more sense to
generalize. Calling on EqSets is now referred to as 'mutation', as what
it's really doing is *mutating* a set to a version of the same thing
which doesn't have all the same values. Since the term 'equivalent' also
seems like it should be used more generally to describe this
'expansion', I'm thinking of just renaiming EqSets to 'Mutate' or
something similar. Anyways, both of these expansion modes have their own
chains of functions now for going from expanding to marking.

The block comments were updated, but the main overview of the program
was not, so I'll do that later. As far as runtime, it was noticeably
impacted, but I think it's still worth it and at some point I'll
probably make a preprocessor macro for turning off input validation
and assuming correctness.