M-Value Ranges for Set Records
Implemented the workings for M-value range segmentation in the backend
of Set Records. Added the necessary fields to the information structure,
and updated the logic of Mark and Query to handle the segmentation. I
did a small test where I replaced a final record with one with a
restricted M-value range, and it seemed to behave correctly.

Still needs to be implemented in the Import/Export functions. Also, to
make this a user-level feature, I'm planning on breaking up the
initialization function in two: one that takes the set size and sets up
the basic info structure (initialize), and one that takes in the range
of M-values and allocates the memory (alloc). Import would take a
structure with the size, and provide the M-value range and allocation
itself. Finally, I'm thinking that the Mark function shouldn't error out
based on M-values, and should just quietly do nothing if the input set
isn't in the allocated range.

Other changes: The helper function for incrementing a set doesn't refer
to M-value or overflowing anymore, as the Query function definitely
knows when to stop. Also, the Query function no longer has the option
for parallelism through segmentation, but I left a bit of commented-out
code in case you want it back. Also there was a memory leak in Query.