Macro for Calculating Record Array Size
I played around with creating some sort of function... at first I wanted
it to allocate the array, then I decided there were some cases
(Import/Export) that would benefit from a simple size calculator, so I
did that, but then I figured I should leave out error checking, and that
made it a one-line function for a simple computation. So now it's just a
macro. It's fine to use macros, and this is what they're intended for.