Import/Export use M-Value Ranges
Implemented the M-Value Range segmentation system into the file I/O.
The Import function now takes any Record of the same set size, but it
overwrites the M-value parameters with those of the file, making a new
array. Tested by modifying the file header of the output from the test
program; it seems to work just as expected.

The new paradigm will be that the M-Value Range is variable, and can be
changed by calling the Allocation function on a record struct. It can
also be changed by importing a Record with a different range. The set
size will remain constant, though, and is only set upon instantiation.