User-Level Set Record Allocation
The Set Record library now requires that a record be manually allocated
through the Allocate function after instantiation. A Record is
instantiated with the Set Size, and the M-Value Range is set by use of
Allocate or Import. The Main and test programs have been updated and
appear to still be working correctly.

Some more things to consider: maybe don't have Mark error out if input
set doesn't conform to M-Value range (or exit differently); go through
the library and make sure the error system is clean; add ability for
user to add bits to header (might want a way to have the program leave a
trace to show all it's done, e.g. has this record been exhaustively
tested?)...