Null Pointers, Function Documentation
Recategorized null pointer checks as input errors. Also now checking for
null record arrays, as the record might've never been allocated. Yeah,
it added some execution time because of all the ridiculous amounts of
validation for every single call to Mark. But I'll fix it for the next
version by enabling that to be cut out with a macro.