Renamed 'Equivalent Sets' to 'Mutate'
The name 'Equivalent Sets' seems like it should be used more generally,
in reference to any means of expanding a set. The name 'Mutate' seems to
fit the method more, just because it's more specific.

Next up, making all the separate programs. Then we'll do the NO_VALIDATE
macro and try to align with the C standard. That should give us a good
2.0.